requirements.txt
fastapi
uvicorn
celery
redis
opencv-python
pillow
requests
fitz
lxml
bs4
numpy
scipy
shapely
ezdxf
rtree
setuptools
PyMuPDF
func_timeout
pytesseract
matplotlib